% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_csvs_from_OSF.R
\name{bind_csvs_from_OSF}
\alias{bind_csvs_from_OSF}
\title{Row‑bind every top‑level CSV file in an OSF component}
\usage{
bind_csvs_from_OSF(
  osf_id,
  pat = Sys.getenv("OSF_PAT"),
  folder = NULL,
  verbose = TRUE
)
}
\arguments{
\item{osf_id}{5‑character GUID or full OSF URL (e.g. "u28tr").}

\item{pat}{Personal‑access‑token (defaults to Sys.getenv("OSF_PAT")).}

\item{folder}{A folder if you want to keep the raw CSV, if NULL then they are deleted}

\item{verbose}{Print progress messages?  Default = TRUE.}
}
\value{
A tibble containing all rows from every *.csv file, padded with NA
for missing columns, plus a \code{.source_file} provenance column.
}
\description{
Row‑bind every top‑level CSV file in an OSF component
}
