% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaforge.R
\name{metaforge_get_token}
\alias{metaforge_get_token}
\title{Get stored MetaForge API Token}
\usage{
metaforge_get_token()
}
\value{
The stored API token or NULL if not set
}
\description{
Get stored MetaForge API Token
}
\keyword{internal}
