% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaforge.R
\docType{package}
\name{MetaForgeR}
\alias{MetaLab}
\alias{MetaLab-package}
\alias{MetaForgeR}
\title{MetaForge API Functions for R}
\description{
Download experiment data from MetaForge (Protocol.js) directly into R.
}
\author{
\strong{Maintainer}: Kit Double \email{kit.double@sydney.edu.au}

}
