% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaforge.R
\name{metaforge_get_all_data}
\alias{metaforge_get_all_data}
\title{Download All MetaForge Project Data}
\usage{
metaforge_get_all_data(token = NULL, combine = FALSE)
}
\arguments{
\item{token}{Optional API token.}

\item{combine}{If TRUE, combines all data into one data frame.}
}
\value{
Either a combined data frame or a named list of data frames
}
\description{
Download All MetaForge Project Data
}
